backend/requirements-minimal.txt
Django==5.1.3
djangorestframework==3.14.0
django-cors-headers==4.2.0
djangorestframework-simplejwt==5.3.1
Pillow==11.3.0
python-dotenv==1.1.1
psycopg2-binary
PyMySQL==1.1.1
requests==2.32.3